added basic caching and refactoring
