+++ flaskr/config.py
@@ -22,6 +22,9 @@ class DevelopmentConfig(Config):
     SQLALCHEMY_DATABASE_URI = environ.get("SQLALCHEMY_DATABASE_URI")
     SQLALCHEMY_TRACK_MODIFICATIONS = False
 
+    CACHE_TYPE = "SimpleCache"
+    CACHE_DEFAULT_TIMEOUT = 300
+
 class TestingConfig(Config):
     TESTING = True
